% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_countymap.R
\name{cp_countymap}
\alias{cp_countymap}
\title{Creates a simple map that highlights a Colorado County}
\usage{
cp_countymap(fips)
}
\arguments{
\item{fips}{is the fips code for the county to highlight}
}
\description{
This function creates a map that highlights one county to be used in the profile process.
}

