% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_ed_short.R
\name{cp_ed_short}
\alias{cp_ed_short}
\title{Creates a Chart comparing educational attainment of two areas}
\usage{
cp_ed_short(fips, state = "08", fips2 = "", state2 = "08", base = 12)
}
\arguments{
\item{fips}{is the fips code for the main area to be compared}

\item{state}{is the state that the original fips}

\item{fips2}{is the second area to be compared Defaults to Blank}

\item{state2}{is the state of the second place to be compared, is set to call up CO since fips2 is blank Defaults to 08}

\item{base}{is the abse text size for the ggplot2 object and codemog_theme()}
}
\description{
Uses the codemog_api function to access ACS data (defaults to 13-5yr) to create a ggplot2 chart for
use in profiles.  This Education Chart shows fewer educational categories.
}

