% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cp_GQage.R
\name{cp_GQage}
\alias{cp_GQage}
\title{Creates a bar chart showing the Group Quarters population by age}
\usage{
cp_GQage(fips, base = 12)
}
\arguments{
\item{fips}{is the fips code for the county to be charted}

\item{base}{is the base text size for the ggplot2 object and codemog_theme()}
}
\description{
This function creates a ggpplot2 chart that shows the GQ population
by age at the 2010 Census.  The chart
gives an idea of how important GQ is for an age group in a place.
}

