% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cp_indchange.R
\name{cp_indchange}
\alias{cp_indchange}
\title{Creates a chart that shows the change in jobs from peak employment to present by industry and CO county}
\usage{
cp_indchange(fips, countyname, base = 12, peakyear = 1)
}
\arguments{
\item{fips}{is the fips code for the main area to be compared}

\item{base}{is the base text size for the ggplot2 object and codemog_theme() Deafults to 12}

\item{peakyear}{is both a switch and an input, if set to 1 it uses the max employment between 2002 and 2008, otherwise it uses what is supplied.}

\item{state}{is the state that the original fips}
}
\description{
This function creates a chart that compares the job counts in an industry at the more recent estimate year to
the year that the county experienced peak employment.  The function does allow for alternative peak years to be
used by specifying the peakyear parameter to a value between 2002 and 2008.
}

