% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_jobShare.R
\name{cp_jobShare}
\alias{cp_jobShare}
\title{Creates a chart that shows the Share of  Employment by industry for CO counties}
\usage{
cp_jobShare(fips, countyname, base = 12)
}
\arguments{
\item{fips}{is the fips code for the main area to be compared}

\item{countyname}{is the name of the county the chart is being made for.}

\item{base}{is the base text size for the ggplot2 object and codemog_theme() Deafults to 12}
}
\description{
This function produces a chart showwing the percentage of the total employment
in each industry.
}

