% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cp_laborforceage.R
\name{cp_laborforceage}
\alias{cp_laborforceage}
\title{Creates a chart showing the laborforce by age for a suppluies CO county compared to the State}
\usage{
cp_laborforceage(fips, countyname, base)
}
\arguments{
\item{fips}{is the fips code for the main area to be compared}

\item{base}{is the base text size for the ggplot2 object and codemog_theme() Deafults to 12}

\item{state}{is the state that the original fips}
}
\description{
This function uses State Demography Office Labor Force by Age Estimates by County to
create a chart that compares the distribution of a given county to the State.
}

