% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cp_laborforcebls.R
\name{cp_laborforcebls}
\alias{cp_laborforcebls}
\title{Creates a timeseries chart of the Labor Force in CO}
\usage{
cp_laborforcebls(fips, countyname, base)
}
\arguments{
\item{fips}{is the fips code for the main area to be compared}

\item{base}{is the base text size for the ggplot2 object and codemog_theme() Deafults to 12}

\item{state}{is the state that the original fips}
}
\description{
This function uses labior force data from the Bureau of Labor Statistics to create a time series of the labor
force for a given CO county.
}

