% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_migbyage.R
\name{cp_migbyage}
\alias{cp_migbyage}
\title{Creates a Chart of Nig Migration Age Profiles}
\usage{
cp_migbyage(fips_list, base = 12, agecat = "five")
}
\arguments{
\item{fips_list}{is the fips code(s) for the countt(ies) of interest}

\item{base}{is the base text size for the ggplot2 object and codemog_theme()}
}
\description{
Uses State Demography Office estimates of the Net Migration by age for each county to
create a chart that shows net migration by single year of age for a list of counties.
Data comes from a demographic analysis of each county using births and deaths from each year
to interpolate the implied migration between the two decennial census' in 2000 and 2010.
}

