% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cp_migjobs.R
\name{cp_migjobs}
\alias{cp_migjobs}
\title{Creates a chart with two plots, one for population trend, one for components of change}
\usage{
cp_migjobs(fips, countyname, base = 12)
}
\arguments{
\item{fips}{is the fips code for the main area to be compared}

\item{base}{is the base text size for the ggplot2 object and codemog_theme()}

\item{state}{is the state that the original fips}
}
\description{
This function creates two seperate time-series plots, a line chart for popultion trends
and a dodged bar chart for components of change over time.  The two plots are stacked together and
placed into one chart.
}

