% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cp_popGQ.R
\name{cp_popGQ}
\alias{cp_popGQ}
\title{Creates an Area chart showing the Group Quarters and Total Population}
\usage{
cp_popGQ(fips, countyname, base = 12)
}
\arguments{
\item{fips}{is the fips code for the main area to be compared}

\item{countyname}{is the name for the county to be used in the title.}

\item{base}{is the base text size for the ggplot2 object and codemog_theme()}
}
\description{
This function creates a ggpplot2 chart that shows the total population
and the group quarter population over time from 1985 to 2013.  The chart
gives an idea of how important GQ is for an area.
}

