% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cp_poptrend.R
\name{cp_poptrend}
\alias{cp_poptrend}
\title{Creates a chart with two plots, one for population trend, one for components of change}
\usage{
cp_poptrend(fips, countyname, base = 12)
}
\arguments{
\item{fips}{is the fips code for the main area to be compared}

\item{countyname}{is the name for the county to be used in the title.}

\item{base}{is the base text size for the ggplot2 object and codemog_theme()}
}
\description{
This function creates two seperate time-series plots, a line chart for popultion trends
and a dodged bar chart for components of change over time.  The two plots are stacked together and
placed into one chart.
}

