% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_census_age.R
\name{ms_census_age}
\alias{ms_census_age}
\title{Creates a chart comparing an areas age distributions in 2000 and 2010}
\usage{
ms_census_age(fips, state = "08", base = 12)
}
\arguments{
\item{fips}{is the fips code for the place or county}

\item{state}{is the state that of the fips in the first argument}
}
\description{
Uses the codemog_api function to access Census data to create a ggplot2 chart for
use in profiles comparing the age distribution at the 2000 and 2010 Censuses.
}

