% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ms_ed.R
\name{ms_ed}
\alias{ms_ed}
\title{Creates a Chart comparing educational attainment of two areas}
\usage{
ms_ed(fips, state = "08", fips2 = "", state2 = "08", base = 12)
}
\arguments{
\item{fips}{is the fips code for the main area to be compared}

\item{state}{is the state that the original fips}

\item{fips2}{is the second area to be compared Defaults to Blank}

\item{state2}{is the state of the second place to be compared, is set to call up CO since fips2 is blank Defaults to 08}

\item{base}{is the abse text size for the ggplot2 object and codemog_theme()}
}
\description{
Uses the codemog_api function to access ACS data (defaults to 13-5yr) to create a ggplot2 chart for
use in profiles.  Included here as more of an example.
}

