% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ms_forecast.R
\name{ms_forecast}
\alias{ms_forecast}
\title{Creates a Chart comparing Forecast Job and Population Growth in CO}
\usage{
ms_forecast(fips, base = 12)
}
\arguments{
\item{fips}{is the fips code for the county being examined}

\item{base}{is the base text size for the ggplot2 object and codemog_theme()}
}
\description{
Uses the data from the State Demography Office package codemog to
create a growth showing population and job changes by Colorado County
from 2010 to 2030.
}

