% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ms_hh.R
\name{ms_hh}
\alias{ms_hh}
\title{CO Household Types Bar Chart}
\usage{
ms_hh(fips, base = 12, state = "08")
}
\arguments{
\item{fips}{The FIPS of the Place or County to use for the graph}

\item{base}{The base font size (in points) to use. Defaults to 12.}

\item{state}{The State FIPS to use.  Defaults to CO.}
}
\description{
This function creates a bar graphs using ggplot2
 that shows household change from 2000 to 2010 by comparing Census Counts.
}

