% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ms_income.R
\name{ms_income}
\alias{ms_income}
\title{Income Distribution Graph}
\usage{
ms_income(fips, fips2 = "", state = "08", state2 = "08")
}
\arguments{
\item{fips}{The FIPS of the Place or County to use}

\item{fips2}{The FIPS of the Place or County to use for comparison}

\item{state}{The State FIPS to use.  Defaults to CO.}

\item{state2}{The State FIPS to use as comparison.  Defaults to CO.}
}
\description{
This function pulls data and generates a graph of the income distribution
 for the areas selected based on the ACS 09-13 5-year File.
}

