% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_jobs.R
\name{ms_jobs}
\alias{ms_jobs}
\title{Creates a Chart showing the Total Estimates Jobs series for each County in Colorado}
\usage{
ms_jobs(fips, countyname, base = 12)
}
\arguments{
\item{fips}{is the fips code for the county being examined}

\item{countyname}{This parameter puts the name of the county in the chart}

\item{base}{is the abse text size for the ggplot2 object and codemog_theme()}
}
\description{
Uses State Demography Office data to create a chart showing the timeseries of Total Estimated Jobs
(which means it includes Proprietors and Agricultural Workers) for each Colorado county in 2013.
}

