% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ms_popage.R
\name{ms_popage}
\alias{ms_popage}
\title{Creates a Chart comparing Forecast Population Growth by Age in Colorado.}
\usage{
ms_popage(fips, base = 12, agegroup = "ten")
}
\arguments{
\item{fips}{is the fips code for the county being examined}

\item{base}{is the base text size for the ggplot2 object and codemog_theme()}
}
\description{
Uses the data from the State Demography Office package codemog to
create a graph showing projected population  changes by Age for each Colorado county from
2015 to 2025.
}

