% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ms_race.R
\name{ms_race}
\alias{ms_race}
\title{Race Function}
\usage{
ms_race(fips, state = "08")
}
\arguments{
\item{fips}{The FIPS of the Place or County to use}

\item{state}{The State FIPS to use.  Defaults to CO.}

\item{fips2}{The FIPS of the Place or County to use for comparison}

\item{state2}{The State FIPS to use as comparison.  Defaults to CO.}
}
\description{
This function pulls data and generates a graph of the Racial Distribution
 for the areas selected based on the 2000 and 2010 Census data files.
}

