% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_pull.R
\name{cp_county}
\alias{cp_county}
\title{MainStreets Pull Function - Counties}
\usage{
cp_county(fips, fips2 = "", state = "08", state2 = "08", od = "")
}
\description{
This function creates the tables, graphs, and file paths for merging
 into the Community Profile InDesign template. The output of the function
 is a dataframe with the proper formatting to do the data merege.
 @param fips The FIPS of the Place or County to use
 @param state  The State FIPS to use.  Defaults to CO.
 @param fips2 The FIPS of the Place or County to use for comparison
 @param state2 The State FIPS to use as comparison.  Defaults to CO.
 @param od The output directory for the graphs being created.
}

