% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/cp_countymap.R
\name{cp_countymap}
\alias{cp_countymap}
\title{Creates a simple map that highlights a Colorado County}
\usage{
cp_countymap(fips)
}
\arguments{
\item{fips}{is the fips code for the county to highlight}
}
\description{
This function creates two seperate time-series plots, a line chart for popultion trends
and a dodged bar chart for components of change over time.  The two plots are stacked together and
placed into one chart.
}

