% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_hh.R
\name{ms_hh}
\alias{ms_hh}
\title{CO Household Types Bar Chart}
\usage{
ms_hh(fips, base = 12, state = "08")
}
\description{
This function creates a bar graphs using ggplot2
 that shows household change from 2000 to 2010 by comparing Census Counts.
 @param fips The FIPS of the Place or County to use for the graph
 @param base The base font size (in points) to use. Defaults to 12.
 @param state  The State FIPS to use.  Defaults to CO.
}

