% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ms_housing.R
\name{ms_housing}
\alias{ms_housing}
\title{CO Housing Unit Table}
\usage{
ms_housing(fips, state = "08")
}
\arguments{
\item{fips}{The FIPS of the Place or County to use for the graph}

\item{state}{The State FIPS to use.  Defaults to CO.}
}
\description{
This function pulls data on housing types by vacancy, vacancy type, and
 tenure from the 2000 and 2010 Census data.
}

