% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_mhi.R
\name{ms_mhi}
\alias{ms_mhi}
\title{Median Household Income Function}
\usage{
ms_mhi(fips, fips2 = "", state = "08", state2 = "08")
}
\description{
This function pulls data and generates a graph of the income distribution
 for the areas selected based on the ACS 09-13 5-year File.
}
\details{
@param fips The FIPS of the Place or County to use
 @param state  The State FIPS to use.  Defaults to CO.
 @param fips2 The FIPS of the Place or County to use for comparison
 @param state2 The State FIPS to use as comparison.  Defaults to CO.
}

