% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_race.R
\name{ms_race}
\alias{ms_race}
\title{Race Function}
\usage{
ms_race(fips, state = "08")
}
\description{
This function pulls data and generates a graph of the Racial Distribution
 for the areas selected based on the 2000 and 2010 Census data files.
}
\details{
@param fips The FIPS of the Place or County to use
 @param state  The State FIPS to use.  Defaults to CO.
 @param fips2 The FIPS of the Place or County to use for comparison
 @param state2 The State FIPS to use as comparison.  Defaults to CO.
}

